Add validator balances to spec.json.
Adding an option to `make_spec.js` in the `pos-contracts` repo to give
initial balances to validators would be a mistake, since a mistake could
allow a network to go into production with the validators having
balances that they should not have.  However, tests require the
validators to have a balance, to pay for gas.  The solution is to
generate a private, ephemeral copy of `spec.json` from the one generated
by the upstream `make_spec.js`.  This is done by
`scripts/add-validator-balances.js`.

Additionally, several cleanups are made to `scripts/getReservedPeer.js`,
and `scripts/start-test-setup` and `package.json` are changed so that
`scripts/add-validator-balances.js` is run with the correct environment
at the correct time.